--warnings
git-svn-id: svn+ssh://forge.imag.fr/var/lib/gforge/chroot/scmrepos/svn/life/trunk/life/trunk@8553 18f2cc81-8059-4896-b63e-f2d89ec8fd72

+++ feel/feelalg/vectorublas.cpp
@@ -647,7 +647,7 @@ VectorUblas<T,Storage>::sqrt() const
     tbb::parallel_for( tbb::blocked_range<size_t>(0, this->localSize() ),
                        detail::Sqrt<this_type>( *this, _tmp ) );
 #else
-    for(int i = 0; i < this->localSize(); ++i )
+    for(int i = 0; i < (int)this->localSize(); ++i )
         _tmp[i] = math::sqrt( this->operator[](i) );
 #endif // HAVE_TBB
 
@@ -659,7 +659,7 @@ typename VectorUblas<T,Storage>::this_type
 VectorUblas<T,Storage>::pow(int n) const
 {
     this_type _out( this->map() );
-    for(int i = 0; i < this->localSize(); ++i )
+    for(int i = 0; i < (int)this->localSize(); ++i )
         _out[i] = math::pow( this->operator[](i), n );
     return _out;
 }
